+++ src/types.ts
@@ -1,0 +1,7 @@
+export enum FileNameFormatters {
+	"pascalCase" = "pascalCase",
+	"camelCase" = "camelCase",
+	"kebabCase" = "kebabCase",
+	"snakeCase" = "snakeCase",
+	"lowerCase" = "lowerCase",
+}